% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_awb_table.R
\name{get_awb_table}
\alias{get_awb_table}
\title{Extracting Auto-White-Balance (AWB) status of images for PhenoCam sites}
\usage{
get_awb_table(site)
}
\arguments{
\item{site}{a character string of the PhenoCam site name}
}
\value{
a data.table of two columns: midday images and their AWB status ('ON', 'OFF', 'UNKOWN', 'LIKELY_YES', 'LIKELY_NO')
}
\description{
Extracting Auto-White-Balance (AWB) status of images for PhenoCam sites
}
\examples{

awb_table <- get_awb_table('dukehw')
head(awb_table)

}
