% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenos.R
\name{get_phenos}
\alias{get_phenos}
\title{Full list of PhenoCam sites and metadata}
\usage{
get_phenos()
}
\value{
a data.table with a list of all the PhenoCam sites and their metadata
}
\description{
Full list of PhenoCam sites and metadata
}
\examples{

phenos <- get_phenos()
head(phenos)

}
