% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rois.R
\name{get_rois}
\alias{get_rois}
\title{Full list of PhenoCam ROI's and metadata}
\usage{
get_rois()
}
\value{
a data.table with a list of all the PhenoCam ROI's and their metadata
}
\description{
Full list of PhenoCam ROI's and metadata
}
\examples{

rois <- get_rois()
head(rois)

}
