% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getphenos.R
\name{getphenos}
\alias{getphenos}
\title{Full list of PhenoCam sites and metadata}
\usage{
getphenos()
}
\value{
a data.table with a list of all the PhenoCam sites and their metadata
}
\description{
Full list of PhenoCam sites and metadata
}
\examples{

phenos <- getphenos()
head(phenos)

}
