% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrois.R
\name{getrois}
\alias{getrois}
\title{Full list of PhenoCam ROI's and metadata}
\usage{
getrois()
}
\value{
a data.table with a list of all the PhenoCam ROI's and their metadata
}
\description{
Full list of PhenoCam ROI's and metadata
}
\examples{

rois <- getrois()
head(rois)

}
